check_init: CheckUpdate: Protect __info_dic
The private attribute `__info_dic` can still be modified
through the "read-only" attribute `info_dic`.

A simple solution is to have the `info_dic` attribute
always return a copy of `__info_dic`. But this means
taking up more memory.

The best way is to override the implementation of OrderedDict.

Test:

In [1]: from check_list import Apktool
In [2]: a = Apktool()
In [3]: a.info_dic['LATEST_VERSION'] is None
Out[3]: True
In [4]: a.info_dic['LATEST_VERSION'] = 'foo'
In [5]: a.info_dic['LATEST_VERSION'] is None
Out[5]: True