main.py: check_one: Properly disable PageCache
`enable_pagecache` is a class attribute of CheckUpdate, and `request_url` is
a class method of CheckUpdate. The `request_url` method only pays attention
to the `enable_pagecache` attribute of the class, not the `enable_pagecache`
attribute of the instance (because `request_url` is a class method, it
cannot get the instance attribute).

Therefore, setting `enable_pagecache` to False for the CheckUpdate instance
here will have no effect.

The correct approach is to directly modify the `enable_pagecache` attribute
of the class (`cls`), but this will have more impact.

So we build a new class here. It inherits from the original class (`cls`),
and has the same name.